feat(openfaas): enable remote building for openfaas modules
Also updated the OpenFaaS components. I'll follow up with another
commit to make the configuration of the OpenFaaS system components
more flexible.

NOTE: This does not enable integration testing for OpenFaaS because
of an issue with deploying the faas-netes Helm chart (see issue #1045)
