+++ garden-service/src/plugins/openfaas/openfaas.ts
@@ -6,116 +6,42 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import dedent = require("dedent")
 import { join } from "path"
-import { resolve as urlResolve } from "url"
 import { ConfigurationError } from "../../exceptions"
-import { PluginContext } from "../../plugin-context"
-import { joiArray, PrimitiveMap, joiProviderName, joi } from "../../config/common"
-import { Module } from "../../types/module"
-import { ServiceStatus, ServiceIngress, Service } from "../../types/service"
-import {
-  ExecModuleSpec,
-  execModuleSpecSchema,
-  ExecTestSpec,
-  testExecModule,
-  getExecModuleBuildStatus,
-} from "../exec"
-import { KubernetesProvider } from "../kubernetes/config"
+import { ServiceStatus, ServiceIngress } from "../../types/service"
+import { testExecModule } from "../exec"
 import { getNamespace, getAppNamespace } from "../kubernetes/namespace"
-import { dumpYaml, findByName } from "../../util/util"
+import { findByName } from "../../util/util"
 import { KubeApi } from "../kubernetes/api"
 import { waitForResources } from "../kubernetes/status/status"
 import { checkWorkloadStatus } from "../kubernetes/status/workload"
-import { CommonServiceSpec } from "../../config/service"
 import { GardenPlugin } from "../../types/plugin/plugin"
-import { Provider, providerConfigBaseSchema, ProviderConfig } from "../../config/provider"
 import { faasCli } from "./faas-cli"
 import { getAllLogs } from "../kubernetes/logs"
-import { LogEntry } from "../../logger/log-entry"
-import { BuildModuleParams } from "../../types/plugin/module/build"
 import { DeployServiceParams } from "../../types/plugin/service/deployService"
 import { GetServiceStatusParams } from "../../types/plugin/service/getServiceStatus"
-import { ConfigureModuleParams, ConfigureModuleResult } from "../../types/plugin/module/configure"
 import { GetServiceLogsParams } from "../../types/plugin/service/getServiceLogs"
 import { DeleteServiceParams } from "../../types/plugin/service/deleteService"
 import { HelmModuleConfig } from "../kubernetes/helm/config"
-import { keyBy, union } from "lodash"
 import { DEFAULT_API_VERSION, STATIC_DIR } from "../../constants"
 import { ExecModuleConfig } from "../exec"
 import { ConfigureProviderParams, ConfigureProviderResult } from "../../types/plugin/provider/configureProvider"
 import { KubernetesDeployment } from "../kubernetes/types"
+import {
+  configSchema,
+  describeType,
+  getK8sProvider,
+  OpenFaasConfig,
+  OpenFaasModule,
+  OpenFaasProvider,
+  OpenFaasPluginContext,
+  OpenFaasService,
+  getServicePath,
+  configureModule,
+} from "./config"
+import { getOpenfaasModuleBuildStatus, buildOpenfaasModule, writeStackFile, stackFilename } from "./build"
 
 const systemDir = join(STATIC_DIR, "openfaas", "system")
-export const stackFilename = "stack.yml"
-
-export interface OpenFaasModuleSpec extends ExecModuleSpec {
-  handler: string
-  image: string
-  lang: string
-}
-
-export const openfaasModuleSpecSchema = execModuleSpecSchema
-  .keys({
-    dependencies: joiArray(joi.string())
-      .description("The names of services/functions that this function depends on at runtime."),
-    handler: joi.string()
-      .default(".")
-      .posixPath({ subPathOnly: true })
-      .description("Specify which directory under the module contains the handler file/function."),
-    image: joi.string()
-      .description("The image name to use for the built OpenFaaS container (defaults to the module name)"),
-    lang: joi.string()
-      .required()
-      .description("The OpenFaaS language template to use to build this function."),
-  })
-  .unknown(false)
-  .description("The module specification for an OpenFaaS module.")
-
-export const openfaasModuleOutputsSchema = joi.object()
-  .keys({
-    endpoint: joi.string()
-      .uri()
-      .required()
-      .description(`The full URL to query this service _from within_ the cluster.`),
-  })
-
-export interface OpenFaasModule extends Module<OpenFaasModuleSpec, CommonServiceSpec, ExecTestSpec> { }
-export type OpenFaasModuleConfig = OpenFaasModule["_ConfigType"]
-export interface OpenFaasService extends Service<OpenFaasModule> { }
-
-export interface OpenFaasConfig extends ProviderConfig {
-  hostname: string
-}
-
-export const configSchema = providerConfigBaseSchema
-  .keys({
-    name: joiProviderName("openfaas"),
-    hostname: joi.string()
-      .hostname()
-      .description(dedent`
-        The hostname to configure for the function gateway.
-        Defaults to the default hostname of the configured Kubernetes provider.
-
-        Important: If you have other types of services, this should be different from their ingress hostnames,
-        or the other services should not expose paths under /function and /system to avoid routing conflicts.`,
-      )
-      .example("functions.mydomain.com"),
-  })
-
-type OpenFaasProvider = Provider<OpenFaasConfig>
-type OpenFaasPluginContext = PluginContext<OpenFaasConfig>
-
-async function describeType() {
-  return {
-    docs: dedent`
-      Deploy [OpenFaaS](https://www.openfaas.com/) functions using Garden. Requires either the \`kubernetes\` or
-      \`local-kubernetes\` provider to be configured. Everything else is installed automatically.
-    `,
-    outputsSchema: openfaasModuleOutputsSchema,
-    schema: openfaasModuleSpecSchema,
-  }
-}
 
 export function gardenPlugin(): GardenPlugin {
   return {
@@ -128,8 +54,8 @@ export function gardenPlugin(): GardenPlugin {
       openfaas: {
         describeType,
         configure: configureModule,
-        getBuildStatus: getExecModuleBuildStatus,
-        build: buildModule,
+        getBuildStatus: getOpenfaasModuleBuildStatus,
+        build: buildOpenfaasModule,
         // TODO: design and implement a proper test flow for openfaas functions
         testModule: testExecModule,
         getServiceStatus,
@@ -217,9 +143,11 @@ async function configureProvider(
       skipDeploy: false,
       tasks: [],
       tests: [],
-      version: "1.7.0",
+      version: "4.4.0",
       releaseName,
       values: {
+        // TODO: allow setting password in provider config
+        basic_auth: false,
         exposeServices: false,
         functionNamespace: namespace,
         ingress: {
@@ -239,7 +167,11 @@ async function configureProvider(
             },
           ],
         },
-        faasnetesd: {
+        // TODO: make this (and more stuff) configurable
+        faasIdler: {
+          create: false,
+        },
+        faasnetes: {
           imagePullPolicy: "IfNotPresent",
         },
         securityContext: false,
@@ -250,56 +182,6 @@ async function configureProvider(
   const moduleConfigs = [systemModule, templateModuleConfig]
 
   return { config, moduleConfigs }
-}
-
-async function configureModule(
-  { ctx, log, moduleConfig }: ConfigureModuleParams<OpenFaasModule>,
-): Promise<ConfigureModuleResult> {
-  moduleConfig.build.dependencies.push({
-    name: "templates",
-    plugin: ctx.provider.name,
-    copy: [{
-      source: "template",
-      target: ".",
-    }],
-  })
-
-  const dependencies = [`${ctx.provider.name}--system`]
-
-  moduleConfig.serviceConfigs = [{
-    dependencies,
-    hotReloadable: false,
-    name: moduleConfig.name,
-    spec: {
-      name: moduleConfig.name,
-      dependencies,
-    },
-  }]
-
-  moduleConfig.testConfigs = moduleConfig.spec.tests.map(t => ({
-    name: t.name,
-    dependencies: union(t.dependencies, dependencies),
-    spec: t,
-    timeout: t.timeout,
-  }))
-
-  moduleConfig.outputs = {
-    endpoint: await getInternalServiceUrl(<OpenFaasPluginContext>ctx, log, moduleConfig),
-  }
-
-  return moduleConfig
-}
-
-async function buildModule({ ctx, log, module }: BuildModuleParams<OpenFaasModule>) {
-  await writeStackFile(<OpenFaasProvider>ctx.provider, module, {})
-
-  const buildLog = await faasCli.stdout({
-    log,
-    cwd: module.buildPath,
-    args: ["build", "-f", stackFilename],
-  })
-
-  return { fresh: true, buildLog }
 }
 
 async function getServiceLogs(params: GetServiceLogsParams<OpenFaasModule>) {
@@ -316,9 +198,10 @@ async function getServiceLogs(params: GetServiceLogsParams<OpenFaasModule>) {
 
 async function deployService(params: DeployServiceParams<OpenFaasModule>): Promise<ServiceStatus> {
   const { ctx, module, service, log, runtimeContext } = params
+  const k8sProvider = getK8sProvider(ctx.provider.dependencies)
 
   // write the stack file again with environment variables
-  await writeStackFile(<OpenFaasProvider>ctx.provider, module, runtimeContext.envVars)
+  await writeStackFile(<OpenFaasProvider>ctx.provider, k8sProvider, module, runtimeContext.envVars)
 
   // use faas-cli to do the deployment
   await faasCli.stdout({
@@ -328,7 +211,6 @@ async function deployService(params: DeployServiceParams<OpenFaasModule>): Promi
   })
 
   // wait until deployment is ready
-  const k8sProvider = getK8sProvider(ctx.provider.dependencies)
   const namespace = await getAppNamespace(ctx, log, k8sProvider)
   const api = await KubeApi.factory(log, k8sProvider.config.context)
   const resources = await getResources(api, service, namespace)
@@ -379,29 +261,6 @@ async function deleteService(params: DeleteServiceParams<OpenFaasModule>): Promi
   return status
 }
 
-async function writeStackFile(
-  provider: OpenFaasProvider, module: OpenFaasModule, envVars: PrimitiveMap,
-) {
-  const image = getImageName(module)
-
-  const stackPath = join(module.buildPath, stackFilename)
-
-  return dumpYaml(stackPath, {
-    provider: {
-      name: "faas",
-      gateway: getExternalGatewayUrl(provider),
-    },
-    functions: {
-      [module.name]: {
-        lang: module.spec.lang,
-        handler: module.spec.handler,
-        image,
-        environment: envVars,
-      },
-    },
-  })
-}
-
 async function getResources(api: KubeApi, service: OpenFaasService, namespace: string) {
   const deployment = await api.apps.readNamespacedDeployment(service.name, namespace)
   return [deployment]
@@ -444,48 +303,4 @@ async function getServiceStatus({ ctx, module, service, log }: GetServiceStatusP
     version,
     ingresses,
   }
-}
-
-function getImageName(module: OpenFaasModule) {
-  return `${module.name || module.spec.image}:${module.version.versionString}`
-}
-
-function getK8sProvider(providers: Provider[]): KubernetesProvider {
-  const providerMap = keyBy(providers, "name")
-  const provider = <KubernetesProvider>(providerMap["local-kubernetes"] || providerMap.kubernetes)
-
-  if (!provider) {
-    throw new ConfigurationError(`openfaas requires a kubernetes (or local-kubernetes) provider to be configured`, {
-      configuredProviders: Object.keys(providers),
-    })
-  }
-
-  return provider
-}
-
-function getServicePath(config: OpenFaasModuleConfig) {
-  return join("/", "function", config.name)
-}
-
-async function getInternalGatewayUrl(ctx: PluginContext<OpenFaasConfig>, log: LogEntry) {
-  const k8sProvider = getK8sProvider(ctx.provider.dependencies)
-  const namespace = await getNamespace({
-    configStore: ctx.configStore,
-    log,
-    projectName: ctx.projectName,
-    provider: k8sProvider,
-    skipCreate: true,
-  })
-  return `http://gateway.${namespace}.svc.cluster.local:8080`
-}
-
-function getExternalGatewayUrl(provider: OpenFaasProvider) {
-  const k8sProvider = getK8sProvider(provider.dependencies)
-  const hostname = provider.config.hostname
-  const ingressPort = k8sProvider.config.ingressHttpPort
-  return `http://${hostname}:${ingressPort}`
-}
-
-async function getInternalServiceUrl(ctx: PluginContext<OpenFaasConfig>, log: LogEntry, config: OpenFaasModuleConfig) {
-  return urlResolve(await getInternalGatewayUrl(ctx, log), getServicePath(config))
 }
